# Defer Sprintf for coroutine status until stack trace generation