## What was changed
Don't call fmt.Sprintf immediately for coroutine status strings, but pass the arguments in pieces and only call it when a stack trace is requested.

## Why?
This was visible in a profile of some worker pods doing heavy replays so I thought I'd try to improve it. It's a small win and does complicate the code slightly, so I'd understand if you decide it's not worth it.

I created a benchmark out of the [schedules replay test](https://github.com/temporalio/temporal/blob/main/service/worker/scheduler/replay_test.go):

```
before:
     310          18965421 ns/op         4362126 B/op      81220 allocs/op
     306          18982254 ns/op         4362069 B/op      81222 allocs/op
     313          18989794 ns/op         4361430 B/op      81215 allocs/op
after:
     308          18604364 ns/op         4344586 B/op      80149 allocs/op
     315          18610687 ns/op         4344066 B/op      80144 allocs/op
     312          18615051 ns/op         4344294 B/op      80146 allocs/op
```
(those are the middle 3 of 13 runs)

So about a 2% improvement in cpu and a tiny bit less memory.

## Checklist

2. How was this tested:
existing unit tests, plus simple benchmark